+++ packages/node/src/subcommands/testing.service.ts
@@ -2,17 +2,20 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Inject, Injectable } from '@nestjs/common';
+import { NestFactory } from '@nestjs/core';
 import { ApiPromise } from '@polkadot/api';
 import {
   NodeConfig,
-  StoreService,
   TestingService as BaseTestingService,
+  NestLogger,
+  TestRunner,
 } from '@subql/node-core';
-import { Sequelize } from '@subql/x-sequelize';
 import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
 import { ApiService } from '../indexer/api.service';
 import { IndexerManager } from '../indexer/indexer.manager';
+import { ProjectService } from '../indexer/project.service';
 import { ApiAt, BlockContent } from '../indexer/types';
+import { TestingModule } from './testing.module';
 
 @Injectable()
 export class TestingService extends BaseTestingService<
@@ -22,30 +25,46 @@ export class TestingService extends BaseTestingService<
   SubqlProjectDs
 > {
   constructor(
-    sequelize: Sequelize,
     nodeConfig: NodeConfig,
-    storeService: StoreService,
     @Inject('ISubqueryProject') project: SubqueryProject,
-    apiService: ApiService,
-    indexerManager: IndexerManager,
   ) {
-    super(
-      sequelize,
-      nodeConfig,
-      storeService,
-      project,
-      apiService,
-      indexerManager,
-    );
+    super(nodeConfig, project);
   }
 
-  async indexBlock(block: BlockContent, handler: string): Promise<void> {
+  async getTestRunner(): Promise<
+    TestRunner<ApiPromise, ApiAt, BlockContent, SubqlProjectDs>
+  > {
+    const testContext = await NestFactory.createApplicationContext(
+      TestingModule,
+      {
+        logger: new NestLogger(),
+      },
+    );
+
+    await testContext.init();
+
+    const projectService: ProjectService = testContext.get(ProjectService);
+    const apiService = testContext.get(ApiService);
+
+    // Initialise async services, we do this here rather than in factories, so we can capture one off events
+    await apiService.init();
+    await projectService.init();
+
+    return testContext.get(TestRunner);
+  }
+
+  async indexBlock(
+    block: BlockContent,
+    handler: string,
+    indexerManager: IndexerManager,
+    apiService: ApiService,
+  ): Promise<void> {
     const runtimeVersion =
-      await this.apiService.unsafeApi.rpc.state.getRuntimeVersion(
+      await apiService.unsafeApi.rpc.state.getRuntimeVersion(
         block.block.block.header.hash,
       );
 
-    await this.indexerManager.indexBlock(
+    await indexerManager.indexBlock(
       block,
       this.getDsWithHandler(handler),
       runtimeVersion,