+++ packages/node-core/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.1.2] - 2023-07-11
 ### Fixed
+- Fix `TestingService` to run tests in seperate application contexts (#1870)
 - Cache race condition when flushing cache and getting data (#1873)
 - Various improvements for POI feature: (#1869)
   - Benchmarking mmr processing